requirements.txt
# Required packages for mem0ai
mem0ai
python-dotenv

# Qdrant dependencies
qdrant-client

# For local embeddings
sentence-transformers